removed dfxml_reader.cpp
