+++ src/Makefile.defs
@@ -3,10 +3,5 @@
 # It is used by programs that wish to include the DFXML_WRITER or the DFXML_READER
 #
 
-DFXML_WRITER = $(DFXML_SRC_DIR)dfxml_writer.h \
-      $(DFXML_SRC_DIR)hash_t.h \
-      $(DFXML_SRC_DIR)cpuid.h
-
-DFXML_READER = $(DFXML_SRC_DIR)dfxml_reader.cpp \
-      $(DFXML_SRC_DIR)dfxml_reader.h \
-      $(DFXML_SRC_DIR)hash_t.h
+DFXML_WRITER = $(DFXML_SRC_DIR)dfxml_writer.h $(DFXML_SRC_DIR)hash_t.h $(DFXML_SRC_DIR)cpuid.h
+DFXML_READER = DFXML_SRC_DIR)dfxml_reader.h $(DFXML_SRC_DIR)hash_t.h
