+++ src/dfxml_demo.cpp
@@ -1,22 +1,25 @@
-#include "dfxml_config.h"
-#include "dfxml_reader.h"
-
-
 /*
  * DFXML demo program.
  *
  * Simson L. Garfinkel
- * Created for bulk_extractor.
- * This file is public domain.
+ *
+ * Revision History:
+ * 2021 - Cleaned up. Added LGPL copyright notice.
+ *
+ * Copyright (C) 2021 Simson L. Garfinkel.
+ *
+ * LICENSE: LGPL Version 3. See COPYING.md for further information.
+ *
  */
+
+
+#include "dfxml_config.h"
+#include "dfxml_reader.h"
 
 void process(dfxml::file_object &fi)
 {
     std::cout << "fi.filename: " << fi.filename() << "\n";
     std::cout << "  pieces: " << fi.byte_runs.size() << "\n";
-    //    for ( auto const &it: fi.byte_runs ){
-    //std::cout << "   " << it  << "\n";
-    //}
 }
